fix(core): a cycle budget is 64 bits wide, as it was before the run loop moved

+++ CassoEmuCore/EmulatorShell.cpp
@@ -9792,7 +9792,7 @@ void EmulatorShell::ExecuteCpuSlices()
         // the guest-time currency the settle pacing is measured in.
         m_clipboardManager->DrainPasteBuffer (sliceTarget);
 
-        sliceActual = m_machine.RunCycles (sliceTarget);
+        sliceActual = static_cast<uint32_t> (m_machine.RunCycles (sliceTarget));
 
         executed += sliceActual;
